Getting rid of tag module.
