+++ src/endpoints/community.rs
@@ -9,6 +9,7 @@ use lemmy_api_common::{
     GetRandomCommunity,
     ListCommunities,
     ListCommunitiesResponse,
+    Tag,
     actions::{
       CreateCommunity,
       FollowCommunity,
@@ -18,7 +19,9 @@ use lemmy_api_common::{
         ApproveCommunityPendingFollower,
         BanFromCommunity,
         BanFromCommunityResponse,
+        CreateCommunityTag,
         DeleteCommunity,
+        DeleteCommunityTag,
         EditCommunity,
         GetCommunityPendingFollowsCount,
         GetCommunityPendingFollowsCountResponse,
@@ -26,6 +29,7 @@ use lemmy_api_common::{
         ListCommunityPendingFollowsResponse,
         RemoveCommunity,
         TransferCommunity,
+        UpdateCommunityTag,
       },
     },
   },
@@ -209,6 +213,32 @@ impl LemmyClient {
         "community/pending_follows/approve",
         data.into(),
       )
+      .await
+  }
+  /// Create a tag for a community you moderate.
+  ///
+  /// HTTP POST /community/tag
+  pub async fn create_community_tag(&self, data: CreateCommunityTag) -> LemmyResult<Tag> {
+    self
+      .make_request(Method::POST, "community/tag", data.into())
+      .await
+  }
+
+  /// Update an existing tag for a community you moderate.
+  ///
+  /// HTTP PUT /community/tag
+  pub async fn update_community_tag(&self, data: UpdateCommunityTag) -> LemmyResult<Tag> {
+    self
+      .make_request(Method::PUT, "community/tag", data.into())
+      .await
+  }
+
+  /// Delete an existing tag for a community you moderate.
+  ///
+  /// HTTP DELETE /community/tag
+  pub async fn delete_community_tag(&self, data: DeleteCommunityTag) -> LemmyResult<Tag> {
+    self
+      .make_request(Method::DELETE, "community/tag", data.into())
       .await
   }
 }
